requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.93.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cuda-bindings==13.2.0
cuda-pathfinder==1.5.2
cuda-toolkit==13.0.2
distro==1.9.0
docstring_parser==0.17.0
edge-tts==7.2.8
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
llvmlite==0.47.0
MarkupSafe==3.0.3
more-itertools==11.0.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
onnxruntime==1.24.4
openai-whisper==20250625
openwakeword==0.4.0
packaging==26.0
propcache==0.4.1
protobuf==7.34.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
python-dotenv==1.2.2
regex==2026.4.4
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
sniffio==1.3.1
sounddevice==0.5.5
sympy==1.14.0
tabulate==0.10.0
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
